# openapi.yaml
openapi: 3.1.0
info:
  title: Pointsman API
  version: 0.1.0
  description: |
    Decision service. A client sends a state; a decision model answers the
    profile's typed questions with probabilities; the profile's policy turns
    the answers into an action.

    In every answer, `p` is the probability of `value`.

    Every request needs an API token: `Authorization: Bearer <token>`. A token
    can be limited to some profiles.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8787
    description: Local development (wrangler dev)
security:
  - bearer: []
paths:
  /v1/decide/{profile}:
    post:
      operationId: decide
      summary: Get a decision from a profile
      parameters:
        - name: profile
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$' }
        - name: version
          in: query
          required: false
          description: Profile version. The latest version when not given.
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecideRequest' }
            example:
              state: { issue: { title: Login page is blank, body: Since this morning the login page shows nothing. } }
              ref: 'example:123'
      responses:
        '200':
          description: The decision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Decision' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: The token may not use this profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/decisions/{id}:
    get:
      operationId: getDecision
      summary: Get a logged decision with its feedback
      parameters:
        - { $ref: '#/components/parameters/DecisionId' }
      responses:
        '200':
          description: The decision as logged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DecisionRecord' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown decision, or one whose profile this token may not use.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/decisions/{id}/feedback:
    post:
      operationId: submitFeedback
      summary: Correct the answers of a decision
      description: |
        Corrections are checked against the profile version the decision used.
        A decision can get feedback more than once; all of it is kept.
      parameters:
        - { $ref: '#/components/parameters/DecisionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeedbackRequest' }
            example:
              correct: { team: frontend }
              by: reviewer@example.com
      responses:
        '204': { description: Feedback stored }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown decision, or one whose profile this token may not use.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: The profile version of the decision is no longer available.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/reviews:
    get:
      operationId: listReviews
      summary: List decisions waiting for a human review
      description: Decisions with action `review` that are not resolved yet, oldest first (at most 100), for the profiles this token may use.
      parameters:
        - name: profile
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Pending reviews
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [reviews]
                properties:
                  reviews:
                    type: array
                    items:
                      allOf: [{ $ref: '#/components/schemas/DecisionRecordBase' }]
                      unevaluatedProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: The token may not use this profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/reviews/{id}/resolve:
    post:
      operationId: resolveReview
      summary: Resolve a pending review
      description: |
        Sets the final action and answers. Answers in `correct` replace the
        model's (source `human`); the others are kept (source `model`).
        `correct` is also stored as feedback. When the decision has a
        `callback_url`, the final answer is sent there (see "Callbacks").
      parameters:
        - { $ref: '#/components/parameters/DecisionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveRequest' }
            example: { action: auto, correct: { team: frontend }, by: reviewer@example.com }
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resolution' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown decision, or one whose profile this token may not use.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: The decision has no pending review (not a review, or already resolved), or its profile version is gone.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/profiles:
    get:
      operationId: listProfiles
      summary: List the profiles this token may use, with their versions
      responses:
        '200':
          description: The profiles this token may use (all profiles for a token with scope *)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [profiles]
                properties:
                  profiles:
                    type: array
                    items: { $ref: '#/components/schemas/ProfileSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/Error' }
components:
  parameters:
    DecisionId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Per-client API token (`pm_` + 43 characters).
  responses:
    Unauthorized:
      description: Missing, malformed, unknown or revoked token.
      headers:
        WWW-Authenticate:
          schema: { type: string, const: Bearer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    DecideRequest:
      type: object
      additionalProperties: false
      required: [state]
      properties:
        state:
          description: |
            What the model decides about: free text, or any JSON object or array.
            When the profile has an `input` mapping, this is the raw client
            payload the mapping reads from.
          oneOf:
            - { type: string, minLength: 1 }
            - { type: object }
            - { type: array }
        ref:
          description: The client's reference for this decision, for example an issue id.
          type: string
          minLength: 1
          maxLength: 200
        callback_url:
          description: Where to send the final answer after a human review (https only).
          type: string
          format: uri
          pattern: '^https://'
    Decision:
      type: object
      additionalProperties: false
      required: [decision_id, answers, action, profile, profile_version, model, created_at, rule]
      properties:
        decision_id: { type: string, format: uuid }
        ref: { type: string }
        answers:
          type: object
          description: One answer per question, keyed by question name.
          additionalProperties: { $ref: '#/components/schemas/Answer' }
        action:
          description: auto, review, or a custom action defined by the profile.
          type: string
        profile: { type: string }
        profile_version: { type: integer, minimum: 1 }
        model:
          description: The model that answered.
          type: string
        created_at: { type: string, format: date-time, description: When the decision was made. }
        rule:
          description: Index of the policy rule that matched (from 0); null when the default action applied.
          type: [integer, 'null']
          minimum: 0
        facts: { $ref: '#/components/schemas/Facts' }
    DecisionRecord:
      allOf:
        - $ref: '#/components/schemas/DecisionRecordBase'
        - type: object
          required: [feedback]
          properties:
            feedback:
              type: array
              items: { $ref: '#/components/schemas/Feedback' }
            callback: { $ref: '#/components/schemas/CallbackState' }
      unevaluatedProperties: false
    DecisionRecordBase:
      type: object
      required: [decision_id, created_at, client, profile, profile_version, model, action, rule, answers, state_hash]
      properties:
        decision_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        client: { type: string, description: The client whose token made the decision. }
        ref: { type: string }
        profile: { type: string }
        profile_version: { type: integer, minimum: 1 }
        model: { type: string }
        action: { type: string }
        rule:
          description: Index of the policy rule that matched; null when the default action applied.
          type: [integer, 'null']
          minimum: 0
        answers:
          type: object
          additionalProperties: { $ref: '#/components/schemas/Answer' }
        facts: { $ref: '#/components/schemas/Facts' }
        state_hash: { type: string, pattern: '^[0-9a-f]{64}$', description: SHA-256 of the state sent to the model (keys sorted). }
        state: { description: 'The state sent to the model. Only when the profile sets log.store_state.' }
        review: { $ref: '#/components/schemas/Review' }
    Review:
      description: Only for decisions with action `review`.
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, enum: [pending, resolved] }
        resolved_at: { type: string, format: date-time }
        resolved_by: { type: string }
        final_action: { type: string }
        final_answers: { $ref: '#/components/schemas/FinalAnswers' }
    FinalAnswers:
      type: object
      additionalProperties:
        type: object
        additionalProperties: false
        required: [value, source]
        properties:
          value: {}
          source: { type: string, enum: [model, human] }
    CallbackState:
      description: Delivery of the final answer to the decision's callback_url.
      type: object
      additionalProperties: false
      required: [status, attempts]
      properties:
        status: { type: string, enum: [pending, delivered, failed] }
        attempts: { type: integer, minimum: 0 }
        last_error: { type: string }
        next_at: { type: string, format: date-time }
    ResolveRequest:
      type: object
      additionalProperties: false
      required: [action, by]
      properties:
        action: { type: string, pattern: '^[a-z][a-z0-9_-]{0,62}$', description: 'The final action, e.g. auto; not review.' }
        correct:
          description: Corrected answers by question name (as in feedback). Empty or left out to keep the model's answers.
          type: object
        by: { type: string, minLength: 1, maxLength: 100 }
        note: { type: string, maxLength: 1000 }
    Resolution:
      type: object
      additionalProperties: false
      required: [decision_id, action, answers, resolved_by, resolved_at, callback]
      properties:
        decision_id: { type: string, format: uuid }
        action: { type: string }
        answers: { $ref: '#/components/schemas/FinalAnswers' }
        resolved_by: { type: string }
        resolved_at: { type: string, format: date-time }
        callback:
          type: string
          enum: [pending, none]
          description: pending when the final answer will be sent to the callback_url.
    FeedbackRequest:
      type: object
      additionalProperties: false
      required: [correct, by]
      properties:
        correct:
          description: 'Corrected values by question name: true/false (noul), an option (choice), a level (score).'
          type: object
          minProperties: 1
          additionalProperties:
            oneOf:
              - { type: boolean }
              - { type: string }
              - { type: integer, minimum: 0 }
        by: { type: string, minLength: 1, maxLength: 100, description: Who made the correction. }
        note: { type: string, maxLength: 1000 }
    Feedback:
      type: object
      additionalProperties: false
      required: [created_at, client, by, correct]
      properties:
        created_at: { type: string, format: date-time }
        client: { type: string }
        by: { type: string }
        correct: { type: object }
        note: { type: string }
    Facts:
      description: >-
        Only for profiles with facts (docs/profile-format.md#facts): each fact the profile asked
        for, by name, as looked up for this decision, or why it is missing. A rule that needs a
        missing fact does not match.
      type: object
      additionalProperties: { $ref: '#/components/schemas/Fact' }
    Fact:
      oneOf:
        - type: object
          additionalProperties: false
          required: [missing, values, source]
          properties:
            missing: { const: false }
            values:
              description: Field values of the fact type (inside, nearest, detour); null when a field has no value.
              type: object
              additionalProperties: { type: [string, number, boolean, 'null'] }
            source: { type: string, description: Where the value came from. }
        - type: object
          additionalProperties: false
          required: [missing, reason]
          properties:
            missing: { const: true }
            reason:
              description: >-
                no_location (no Point or LineString at the fact's path), unavailable (no fact
                provider configured), timeout, or error.
              enum: [no_location, unavailable, timeout, error]
    Answer:
      oneOf:
        - $ref: '#/components/schemas/NoulAnswer'
        - $ref: '#/components/schemas/ChoiceAnswer'
        - $ref: '#/components/schemas/ScoreAnswer'
    NoulAnswer:
      type: object
      additionalProperties: false
      required: [type, value, p, 'yes']
      properties:
        type: { const: noul }
        value: { type: boolean, description: 'true when P(yes) >= 0.5' }
        p: { $ref: '#/components/schemas/Probability' }
        'yes': { $ref: '#/components/schemas/Probability' }
    ChoiceAnswer:
      type: object
      additionalProperties: false
      required: [type, value, p, probabilities]
      properties:
        type: { const: choice }
        value: { type: string, description: The most likely option. }
        p: { $ref: '#/components/schemas/Probability' }
        probabilities: { $ref: '#/components/schemas/Probabilities' }
    ScoreAnswer:
      type: object
      additionalProperties: false
      required: [type, value, p, score, probabilities]
      properties:
        type: { const: score }
        value: { type: integer, minimum: 0, description: The most likely level, 0 = lowest. }
        p: { $ref: '#/components/schemas/Probability' }
        score: { type: number, minimum: 0, description: Probability-weighted level; can fall between levels. }
        probabilities:
          $ref: '#/components/schemas/Probabilities'
          description: Probability per level, keyed by level index.
    Probability: { type: number, minimum: 0, maximum: 1 }
    Probabilities:
      type: object
      additionalProperties: { $ref: '#/components/schemas/Probability' }
    ProfileSummary:
      type: object
      additionalProperties: false
      required: [id, version, versions, title, description]
      properties:
        id: { type: string }
        version: { type: integer, minimum: 1, description: Latest version. }
        versions: { type: array, items: { type: integer, minimum: 1 } }
        title: { $ref: '#/components/schemas/BilingualText' }
        description: { $ref: '#/components/schemas/BilingualText' }
    BilingualText:
      type: object
      additionalProperties: false
      required: [en, ja]
      properties:
        en: { type: string }
        ja: { type: string }
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, unauthorized, forbidden, profile_not_found, decision_not_found, profile_version_gone, not_pending, not_found, model_error, internal_error]
            message: { type: string }